requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dateutil==2.8.2
requests==2.31.0
gunicorn==21.2.0
